Conflicts:
	lucene-5205/src/main/java/org/apache/lucene/queryparser/spans/SpanQueryParserBase.java
	lucene-5205/src/test/java/org/apache/lucene/queryparser/spans/TestSpanOnlyQueryParser.java

+++ lucene-5317/src/main/java/org/apache/lucene/corpus/stats/TermDFTF.java
@@ -17,18 +17,17 @@ package org.apache.lucene.corpus.stats;
  */
 
 
-
 public class TermDFTF extends TermDF {
 
-  public final long termFreq;
+  public final int termFreq;
 
-  public TermDFTF(String term, int docFreq, long termFreq) {
+  public TermDFTF(String term, int docFreq, int termFreq) {
     super(term, docFreq);
     this.termFreq = termFreq;
   }
 
 
-  public long getTermFreq() {
+  public int getTermFreq() {
     return termFreq;
   }
 
@@ -42,15 +41,14 @@ public class TermDFTF extends TermDF {
     } else if (docFreq > other.docFreq) {
       return -1;
     }
-    
+
     if (termFreq < other.termFreq) {
       return 1;
     } else if (termFreq > other.termFreq) {
       return -1;
-    }    
+    }
     return term.compareTo(other.getTerm());
   }
-
 
 
   @Override
@@ -83,7 +81,7 @@ public class TermDFTF extends TermDF {
 
 
   @Override
-  public String toString(){
+  public String toString() {
     StringBuilder sb = new StringBuilder();
     sb.append(term).append(": tf=").append(termFreq).append(" df=").append(docFreq);
     return sb.toString();
